requirements/development.txt
-r common.txt

Pillow==1.7.8
glue==0.3
ipython==1.1.0
django-debug-toolbar==0.9.4

flake8
